Inscription et connexion

+++ filRouge/src/main/java/com/filRouge/repository/DemandeServiceRepository.java
@@ -3,5 +3,5 @@ package com.filRouge.repository;
 import com.filRouge.model.DemandeService;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DemandeServiceRepository extends JpaRepository<DemandeService, Integer> {
+public interface DemandeServiceRepository extends JpaRepository<DemandeService, Long> {
 }
